Prefer EDH over EECDH
Default ephemeral Diffie-Hellman is a little slower, but very well
tested.   Elliptic curve DH comes with some concerns regarding the
involvement of NIST, as some NIST recommendations have been shown to
contain cryptographic backdoors.  The ECDH curves are expected to be
secure and not backdoored, so using them seems like an acceptable risk.
Still, EDH is preferred here as a basic precaution.